copied from [nach2]: remove number: true from addendaSequenceNumber
https://github.com/sizzers/nACH2/commit/18444a4069d1e8ea20b9a2321d755bc41bf10dbf

it is numeric, not a number.
all numbers are considered floats here and this is just a sequence number going from 1 to 9999
setting it as number caused two 0 to be added at the end. so 1 became 0100 instead of 0001

P.S: there is an error in the vlidation, numbers are checked against their current size ignoring
the added padding if needed. so a 4 width number for example of value 1234 will pass the validation
but will be resulted as 123400 which is longer than allowed